backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
motor>=3.6.0
beanie>=1.27.0
pydantic>=2.10.3
pydantic-settings>=2.6.1
python-jose[cryptography]>=3.3.0
bcrypt==4.2.1
python-multipart>=0.0.20
httpx>=0.28.1
tenacity>=9.0.0

# SIGMA conversion  (pySigma 1.x + compatible elasticsearch backend 2.x)
pySigma==1.2.0
pySigma-backend-elasticsearch==2.0.2
pyyaml==6.0.3

# AI providers
google-generativeai==0.8.3
openai==1.57.4

# Elasticsearch
elasticsearch[async]==8.16.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
